vp source helper now depends on timeline points
