+++ client/src/SessionView/Overview/VictoryPoints/PointsSourceHelper/index.js
@@ -1,0 +1,1 @@
+export * from './PointsSourceHelper'
